// app/commons/CallToAction/types.ts
export type CtaColor =
  | 'default'
  | 'primary'
  | 'secondary'
  | 'secondaryNeutral'
  | 'toolbarSubmenu'
  | 'infobar'
  | 'navbar'
  | 'error'
  | 'sidebar'
  | 'defaultNeutral'
  | 'field'
  | 'onetPremium';

export type CtaSize = 'xs' | 'sm' | 'md' | 'lg';

export type CtaShape = 'full' | 'square' | 'circle';

export type CtaAlign = 'left' | 'center' | 'right';